// src/app/core/enums/const.enum.ts
export enum ConstEnum {
  token = 'pluto-token',
  username = 'pluto-username',
  password = 'pluto-password',
  language = 'pluto-language',
  selectLangFlag = 'pluto-select-lang-flag',
  layout = 'pluto-layout',
  collapseSidebar = 'pluto-collapse-sidebar',
  dark = 'pluto-dark',
  sidebarColor = 'pluto-sidebar-color',
  headerColor = 'pluto-header-color'
}